+++ node/src/main/kotlin/net/corda/node/internal/Node.kt
@@ -132,6 +132,7 @@ open class Node(configuration: NodeConfiguration,
     override lateinit var serverThread: AffinityExecutor.ServiceAffinityExecutor
 
     private var messageBroker: ArtemisMessagingServer? = null
+    private var bridgeControlListener: BridgeControlListener? = null
 
     private var shutdownHook: ShutdownHook? = null
 
@@ -145,6 +146,7 @@ open class Node(configuration: NodeConfiguration,
 
         val serverAddress = configuration.messagingServerAddress ?: makeLocalMessageBroker()
         val advertisedAddress = info.addresses.single()
+        bridgeControlListener = BridgeControlListener(configuration, serverAddress, networkParameters.maxMessageSize)
 
         printBasicNodeInfo("Incoming connection address", advertisedAddress.toString())
         rpcMessagingClient = RPCMessagingClient(configuration, serverAddress, networkParameters.maxMessageSize)
@@ -162,13 +164,14 @@ open class Node(configuration: NodeConfiguration,
                 serviceIdentity,
                 serverThread,
                 database,
+                services.networkMapCache,
                 advertisedAddress,
                 networkParameters.maxMessageSize)
     }
 
     private fun makeLocalMessageBroker(): NetworkHostAndPort {
         with(configuration) {
-            messageBroker = ArtemisMessagingServer(this, p2pAddress.port, rpcAddress?.port, services.networkMapCache, securityManager, networkParameters.maxMessageSize)
+            messageBroker = ArtemisMessagingServer(this, p2pAddress.port, rpcAddress?.port, securityManager, networkParameters.maxMessageSize)
             return NetworkHostAndPort("localhost", p2pAddress.port)
         }
     }
@@ -220,6 +223,11 @@ open class Node(configuration: NodeConfiguration,
     override fun startMessagingService(rpcOps: RPCOps) {
         // Start up the embedded MQ server
         messageBroker?.apply {
+            runOnStop += this::stop
+            start()
+        }
+        // Start P2P bridge service
+        bridgeControlListener?.apply {
             runOnStop += this::stop
             start()
         }