+++ node/src/main/kotlin/net/corda/node/services/messaging/P2PMessagingClient.kt
@@ -1,10 +1,13 @@
 package net.corda.node.services.messaging
 
+import net.corda.core.crypto.toStringShort
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.internal.ThreadBox
 import net.corda.core.messaging.CordaRPCOps
 import net.corda.core.messaging.MessageRecipients
 import net.corda.core.messaging.SingleMessageRecipient
+import net.corda.core.node.NodeInfo
+import net.corda.core.node.services.NetworkMapCache
 import net.corda.core.node.services.PartyInfo
 import net.corda.core.serialization.SerializationDefaults
 import net.corda.core.serialization.SingletonSerializeAsToken
@@ -15,12 +18,18 @@ import net.corda.core.utilities.contextLogger
 import net.corda.core.utilities.sequence
 import net.corda.core.utilities.trace
 import net.corda.node.VersionInfo
+import net.corda.node.services.api.NetworkMapCacheInternal
 import net.corda.node.services.config.NodeConfiguration
 import net.corda.node.services.statemachine.StateMachineManagerImpl
 import net.corda.node.utilities.AffinityExecutor
 import net.corda.node.utilities.AppendOnlyPersistentMap
 import net.corda.node.utilities.PersistentMap
 import net.corda.nodeapi.internal.ArtemisMessagingComponent.*
+import net.corda.nodeapi.internal.ArtemisMessagingComponent.Companion.BRIDGE_CONTROL
+import net.corda.nodeapi.internal.ArtemisMessagingComponent.Companion.BRIDGE_NOTIFY
+import net.corda.nodeapi.internal.ArtemisMessagingComponent.Companion.PEERS_PREFIX
+import net.corda.nodeapi.internal.BridgeControl
+import net.corda.nodeapi.internal.BridgeEntry
 import net.corda.nodeapi.internal.persistence.CordaPersistence
 import net.corda.nodeapi.internal.persistence.NODE_DATABASE_PREFIX
 import org.apache.activemq.artemis.api.core.ActiveMQObjectClosedException
@@ -29,6 +38,8 @@ import org.apache.activemq.artemis.api.core.RoutingType
 import org.apache.activemq.artemis.api.core.SimpleString
 import org.apache.activemq.artemis.api.core.client.ClientConsumer
 import org.apache.activemq.artemis.api.core.client.ClientMessage
+import org.apache.activemq.artemis.api.core.client.ClientSession
+import rx.Subscription
 import java.security.PublicKey
 import java.time.Instant
 import java.util.*
@@ -74,6 +85,7 @@ class P2PMessagingClient(config: NodeConfiguration,
                          private val serviceIdentity: PublicKey?,
                          private val nodeExecutor: AffinityExecutor.ServiceAffinityExecutor,
                          private val database: CordaPersistence,
+                         private val networkMap: NetworkMapCacheInternal,
                          advertisedAddress: NetworkHostAndPort = serverAddress,
                          maxMessageSize: Int
 ) : SingletonSerializeAsToken(), MessagingService {
@@ -135,6 +147,8 @@ class P2PMessagingClient(config: NodeConfiguration,
         var running = false
         var p2pConsumer: ClientConsumer? = null
         var serviceConsumer: ClientConsumer? = null
+        var bridgeNotifyConsumer: ClientConsumer? = null
+        var networkChangeSubscription: Subscription? = null
     }
 
     private val messagesToRedeliver = database.transaction {
@@ -207,9 +221,93 @@ class P2PMessagingClient(config: NodeConfiguration,
                     }
                 }
             }
+            registerBridgeControl(session, inbox)
+            enumerateBridges(session, inbox)
         }
 
         resumeMessageRedelivery()
+    }
+
+    private fun InnerState.registerBridgeControl(session: ClientSession, inbox: String) {
+        val bridgeNotifyQueue = "$BRIDGE_NOTIFY.${myIdentity.toStringShort()}"
+        session.createTemporaryQueue(BRIDGE_NOTIFY, RoutingType.MULTICAST, bridgeNotifyQueue)
+        val bridgeConsumer = session.createConsumer(bridgeNotifyQueue)
+        bridgeNotifyConsumer = bridgeConsumer
+        bridgeConsumer.setMessageHandler { msg ->
+            val data: ByteArray = ByteArray(msg.bodySize).apply { msg.bodyBuffer.readBytes(this) }
+            val notifyMessage = data.deserialize<BridgeControl>(context = SerializationDefaults.P2P_CONTEXT)
+            log.info(notifyMessage.toString())
+            if (notifyMessage is BridgeControl.BridgeToNodeQuery) {
+                enumerateBridges(session, inbox)
+            }
+            msg.acknowledge()
+        }
+        networkChangeSubscription = networkMap.changed.subscribe { updateBridgesOnNetworkChange(it) }
+    }
+
+    private fun sendBridgeControl(message: BridgeControl) {
+        val client = artemis.started!!
+        val controlPacket = message.serialize(context = SerializationDefaults.P2P_CONTEXT).bytes
+        val artemisMessage = client.session.createMessage(false)
+        artemisMessage.writeBodyBufferBytes(controlPacket)
+        client.producer.send(BRIDGE_CONTROL, artemisMessage)
+    }
+
+    private fun updateBridgesOnNetworkChange(change: NetworkMapCache.MapChange) {
+        log.info("Updating bridges on network map change: ${change.node}")
+        fun gatherAddresses(node: NodeInfo): Sequence<BridgeEntry> {
+            val address = node.addresses.first()
+            return node.legalIdentitiesAndCerts.map {
+                val messagingAddress = NodeAddress(it.party.owningKey, address)
+                BridgeEntry(messagingAddress.queueName, listOf(messagingAddress.hostAndPort), listOf(it.party.name))
+            }.filter { artemis.started!!.session.queueQuery(SimpleString(it.queueName)).isExists }.asSequence()
+        }
+
+        fun deployBridges(node: NodeInfo) {
+            gatherAddresses(node)
+                    .forEach {
+                        sendBridgeControl(BridgeControl.BridgeCreate(it))
+                    }
+        }
+
+        fun destroyBridges(node: NodeInfo) {
+            gatherAddresses(node)
+                    .forEach {
+                        sendBridgeControl(BridgeControl.BridgeDelete(it))
+                    }
+        }
+
+        when (change) {
+            is NetworkMapCache.MapChange.Added -> {
+                deployBridges(change.node)
+            }
+            is NetworkMapCache.MapChange.Removed -> {
+                destroyBridges(change.node)
+            }
+            is NetworkMapCache.MapChange.Modified -> {
+                destroyBridges(change.previousNode)
+                deployBridges(change.node)
+            }
+        }
+    }
+
+    private fun enumerateBridges(session: ClientSession, inbox: String) {
+        val requiredBridges = mutableListOf<BridgeEntry>()
+        fun createBridgeEntry(queueName: SimpleString) {
+            val keyHash = queueName.substring(PEERS_PREFIX.length)
+            val peers = networkMap.getNodesByOwningKeyIndex(keyHash)
+            for (node in peers) {
+                val bridge = BridgeEntry(queueName.toString(), node.addresses, node.legalIdentities.map { it.name })
+                requiredBridges += bridge
+            }
+        }
+
+        val queues = session.addressQuery(SimpleString("$PEERS_PREFIX#")).queueNames
+        for (queue in queues) {
+            createBridgeEntry(queue)
+        }
+        val startupMessage = BridgeControl.NodeToBridgeSnapshot(listOf(inbox), requiredBridges)
+        sendBridgeControl(startupMessage)
     }
 
     private fun resumeMessageRedelivery() {
@@ -361,9 +459,15 @@ class P2PMessagingClient(config: NodeConfiguration,
             check(artemis.started != null)
             val prevRunning = running
             running = false
+            networkChangeSubscription?.unsubscribe()
             val c = p2pConsumer ?: throw IllegalStateException("stop can't be called twice")
             try {
                 c.close()
+            } catch (e: ActiveMQObjectClosedException) {
+                // Ignore it: this can happen if the server has gone away before we do.
+            }
+            try {
+                bridgeNotifyConsumer!!.close()
             } catch (e: ActiveMQObjectClosedException) {
                 // Ignore it: this can happen if the server has gone away before we do.
             }
@@ -493,6 +597,17 @@ class P2PMessagingClient(config: NodeConfiguration,
             if (!queueQuery.isExists) {
                 log.info("Create fresh queue $queueName bound on same address")
                 session.createQueue(queueName, RoutingType.ANYCAST, queueName, true)
+                if (queueName.startsWith(PEERS_PREFIX)) {
+                    val keyHash = queueName.substring(PEERS_PREFIX.length)
+                    val peers = networkMap.getNodesByOwningKeyIndex(keyHash)
+                    for (node in peers) {
+                        val bridge = BridgeEntry(queueName, node.addresses, node.legalIdentities.map { it.name })
+                        val createBridgeMessage = BridgeControl.BridgeCreate(bridge)
+                        val artemisCreateBridgeMessage = session.createMessage(false)
+                        artemisCreateBridgeMessage.writeBodyBufferBytes(createBridgeMessage.serialize(context = SerializationDefaults.P2P_CONTEXT).bytes)
+                        artemis.started!!.producer.send(BRIDGE_CONTROL, artemisCreateBridgeMessage)
+                    }
+                }
             }
         }
     }